Seperated AnimalDiet
