--- student-projects/gilad/vlad-notebook/sessions-12-19-typescript/session-18/home-project/Animal/AnimalType.ts
@@ -1,35 +1,0 @@
-// enum Gender {
-// 	Female,
-// 	Male,
-// }
-
-// class AnimalType {
-// 	constructor(
-// 		private name: string,
-// 		private animalDiet: AnimalDiet,
-// 		private sound: string
-// 	) {
-// 		this.name = name;
-// 		this.animalDiet = animalDiet;
-// 		this.sound = sound;
-// 	}
-
-// 	getName() {
-// 		return this.name;
-// 	}
-
-// 	getDiet() {
-// 		return this.animalDiet;
-// 	}
-
-// 	getSound() {
-// 		return this.sound;
-// 	}
-
-// 	isAgressive() {
-// 		return this.animalDiet.diet === Diets.Herbivore ? false : false;
-// 	}
-// }
-
-// const camel = new AnimalType("Camel", dietHerbivore, "grunt");
-// const bear = new AnimalType("Bear", dietCarnivore, "growl");
