+++ student-projects/gilad/vlad-notebook/sessions-12-19-typescript/session-18/home-project/index.html
@@ -9,7 +9,7 @@
 
 <body>
 
-    <script src="dist/app.js"></script>
+    <script type="module" src="dist/app.js"></script>
 </body>
 
 </html>
